+++ web/src/main.tsx
@@ -2,13 +2,14 @@ import { createTheme, MantineProvider } from "@mantine/core";
 import "@mantine/core/styles.css";
 import "@mantine/notifications/styles.css";
 import "./nucleus";
+import { Notifications } from "@mantine/notifications";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { RouterProvider } from "react-router-dom";
 import ConfigProvider from "./providers/ConfigProvider.tsx";
 import { router } from "./router";
-import { Notifications } from "@mantine/notifications";
+import { HttpError } from "./util/Errors";
 
 const theme = createTheme({
   fontFamily: "Open Sans, sans-serif",
@@ -43,9 +44,26 @@ const theme = createTheme({
   },
 });
 
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      retry: (failureCount, error: Error) => {
+        if (failureCount >= 5 && error instanceof HttpError) {
+          if (error.status === 401) {
+            window.location.href = "/login";
+            return false;
+          }
+        }
+
+        return failureCount < 5;
+      },
+    },
+  },
+});
+
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
-    <QueryClientProvider client={new QueryClient()}>
+    <QueryClientProvider client={queryClient}>
       <MantineProvider theme={theme} forceColorScheme={"dark"}>
         <ConfigProvider>
           <RouterProvider router={router} />